backend/requirements.txt
fastapi
uvicorn
kafka-python
python-dotenv
pandas
numpy
scikit-learn
tensorflow
matplotlib
seaborn
joblib
plotly
shap
requests


